split off numeric_utils header
