+++ include/ck/utility/dtype_fp64.hpp
@@ -4,4 +4,4 @@ namespace ck {
 // fp64
 using double2_t = typename vector_type<double, 2>::type;
 using double4_t = typename vector_type<double, 4>::type;
-}
+} // namespace ck
