+++ include/ck/utility/e8m0.hpp
@@ -64,49 +64,6 @@ struct e8m0_bexp_t
     __host__ __device__ constexpr bool is_nan() const { return data == nan_mask; }
 };
 
-template <>
-struct NumericLimits<e8m0_bexp_t>
-{
-    static constexpr e8m0_bexp_t binary_min  = 0x00; // 0b00000000
-    static constexpr e8m0_bexp_t binary_max  = 0xFE; // 0b11111110
-    static constexpr e8m0_bexp_t binary_qnan = 0xFF; // 0b11111111
-    static constexpr e8m0_bexp_t binary_1    = 0x7F; // 0b01111111
-    static constexpr e8m0_bexp_t binary_2    = 0x80; // 0b10000000
-    static constexpr e8m0_bexp_t binary_3    = 0x82; // 0b10000010
-    static constexpr e8m0_bexp_t binary_135  = 0x87; // 0b10000111
-    static constexpr e8m0_bexp_t binary_142  = 0x8E; // 0b10001110
-
-    __host__ __device__ static constexpr e8m0_bexp_t Min() { return e8m0_bexp_t(binary_min); }
-    __host__ __device__ static constexpr e8m0_bexp_t Max() { return e8m0_bexp_t(binary_max); }
-    __host__ __device__ static constexpr e8m0_bexp_t QuietNaN() { return e8m0_bexp_t(binary_qnan); }
-    __host__ __device__ static constexpr e8m0_bexp_t Binary_1() { return e8m0_bexp_t(binary_1); }
-    __host__ __device__ static constexpr e8m0_bexp_t Binary_2() { return e8m0_bexp_t(binary_2); }
-    __host__ __device__ static constexpr e8m0_bexp_t Binary_3() { return e8m0_bexp_t(binary_3); }
-    __host__ __device__ static constexpr e8m0_bexp_t Binary_135()
-    {
-        return e8m0_bexp_t(binary_135);
-    }
-    __host__ __device__ static constexpr e8m0_bexp_t Binary_142()
-    {
-        return e8m0_bexp_t(binary_142);
-    }
-};
-
-template <>
-struct NumericUtils<e8m0_bexp_t>
-{
-    static constexpr int exp  = 8;
-    static constexpr int mant = 0;
-    static constexpr int bias = 127;
-
-    static constexpr int unbiased_exp_min = -127;
-    static constexpr int unbiased_exp_max = 127;
-    static constexpr int biased_exp_min   = 0;
-    static constexpr int biased_exp_max   = 254;
-
-    using bitwise_type = uint8_t;
-};
-
 namespace utils {
 
 template <typename T>
