+++ include/ck/utility/f8_utils.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "ck/utility/data_type.hpp"
+#include "ck/utility/numerical_utils.hpp"
 
 namespace ck {
